feat(web-domain-data): remove modal link
ref: #DCE-119
