+++ packages/manager/apps/web-domains/src/domain/routes/routes.tsx
@@ -46,10 +46,6 @@ const DsRecordListingPage = React.lazy(() =>
   import('@/domain/pages/domainTabs/dsRecords/dsRecordsListing'),
 );
 
-const DsRecordsDelete = React.lazy(() =>
-  import('@/domain/pages/domainTabs/dsRecords/dsRecordsDelete'),
-);
-
 function RedirectToDefaultTab() {
   const { serviceName } = useParams<{ serviceName: string }>();
   return (
@@ -114,12 +110,7 @@ export default (
             Component={HostConfigurationDeletePage}
           />
         </Route>
-        <Route path={urls.domainTabDsrecords} Component={DsRecordListingPage}>
-          <Route
-            path={urls.domainTabDsrecordsDelete}
-            Component={DsRecordsDelete}
-          />
-        </Route>
+        <Route path={urls.domainTabDsrecords} Component={DsRecordListingPage} />
         <Route
           path={urls.domainTabContactManagement}
           Component={ContactManagementPage}
